pyproject.toml
[project]
name = "authexample"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexander Petrov", email = "101967173+AlexXxShurik@users.noreply.github.com"}
]
requires-python = ">=3.12,<4.0"

dependencies = [
    "uvicorn[standard] >=0.35.0,<0.36.0",
    "fastapi >=0.116.1,<0.117.0",
    "sqlalchemy >=2.0.43,<3.0.0",
    "alembic >=1.16.5,<2.0.0",
    "psycopg2-binary >=2.9.10,<3.0.0",
    "pydantic >=2.11.7,<3.0.0",
    "pydantic-settings >=2.10.1,<3.0.0",
    "colorlog >=6.9.0,<7.0.0",
    "slowapi >=0.1.9,<0.2.0",
    "email-validator >=2.3.0,<3.0.0",
    "python-jose >=3.5.0,<4.0.0",
    "asyncpg >=0.30.0,<0.31.0",
    "greenlet >=3.2.4,<4.0.0",
    "jwt >=1.4.0,<2.0.0",
    "passlib[bcrypt] >=1.7.4,<2.0.0",
    "jinja2 >=3.1.6,<4.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.poetry]
package-mode = false
